continue to refactor benchmark code
