+++ benchmark/test_pe_benchmark_with_llm.py
@@ -3,6 +3,8 @@ import os
 from datetime import datetime
 from string import Template
 
+from benchmark.comm_llm import run_llm_benchmark
+
 from .utils import generate_columns_definition
 from .common import (
     ensure_target_result_directory_existed, 
@@ -14,33 +16,6 @@ from .constant import (
     BenchmarkType,
     LLModelType
 )
-
-system_prompts_template = Template("""
-Please act as a biomedial expert to assess the similarities and differences between two biomedical tables, one is baseline table, the other is extracted table. 
-Provide a similarity rating on a scale of 0 to 100, where 100 indicates identical tables and 0 indicates completely different tables.
-The tables contains the following columns:
-$columns_definition
-
-Note:  Evaluate the similarity of the two tables regardless of their row order and column order. And output similarity score in the format [[{score}]].
-
-""")
-
-table_prompt_template = Template("""
-baseline table :
-$table_baseline
-
-extracted table:
-$table_generated
-Please assess the above two tables using baseline table as the standard.
-Note:
-1. Please use the Parameter type -Value-Unit-P Value-Outcomes-Exposure as an anchor to assess the similarity.
-2.If there are missing rows, you have to reduce the similarity score significantly based on the number of missing rows. e.g., the median and mean are different.
-3. Cells with NaN values will be considered as a blank cell.
-4. You should specify the difference between the two tables.
-5. Synonyms may exist in the tables, e.g., population, drug name, specimen, etc.
-6. The data should be mapped to the correct columns. If the cell contents do not comply with the column definition, this indicates the data is placed into the wrong cells. You should reduce the similarity score. For example, based on the column definition, summary statistics should not contain variation information.
-7. Ignore the lowercase and uppercase letters
-""")
 
 @pytest.mark.skip("just for test the feasible of claude api")
 def test_claude(client):
@@ -75,58 +50,20 @@ def ensured_result_path():
     return os.path.join(result_dir, "result.log")
 
 def test_gpt4o_benchmark(client, prepared_dataset, ensured_result_path):
-    for id in prepared_dataset:
-        the_dict = prepared_dataset[id]
-        baseline = the_dict["baseline"]
-        if not LLModelType.GPT4O.value in the_dict:
-            continue
-        gpt4o = the_dict[LLModelType.GPT4O.value]
-        with open(baseline, "r") as fobj:
-            table_baseline = fobj.read()
-        with open(gpt4o, "r") as fobj:
-            table_gpt4o = fobj.read()
-        user_message = user_message = table_prompt_template.substitute({
-            "table_baseline": table_baseline,
-            "table_generated": table_gpt4o,
-        })
-        cols_definition = generate_columns_definition(BenchmarkType.PE)
-        system_prompts = system_prompts_template.substitute({
-            "columns_definition": cols_definition
-        })
-        msg, usage = client.create(system_prompts,user_message)
-        write_LLM_score(
-            output_fn=ensured_result_path,
-            model=LLModelType.GPT4O.value,
-            pmid=id,
-            score=msg,
-            token_usage = usage,
-        )
+    run_llm_benchmark(
+        dataset=prepared_dataset,
+        benchmark_type=BenchmarkType.PE,
+        model=LLModelType.GPT4O,
+        result_file=ensured_result_path,
+        client=client,
+    )
 
 def test_gemini15_benchmark(client, prepared_dataset, ensured_result_path):
-    for id in prepared_dataset:
-        the_dict = prepared_dataset[id]
-        baseline = the_dict["baseline"]
-        if not LLModelType.GEMINI15.value in the_dict:
-            continue
-        gemini15 = the_dict[LLModelType.GEMINI15.value]
-        with open(baseline, "r") as fobj:
-            table_baseline = fobj.read()
-        with open(gemini15, "r") as fobj:
-            table_gemini15 = fobj.read()
-        user_message = user_message = table_prompt_template.substitute({
-            "table_baseline": table_baseline,
-            "table_generated": table_gemini15,
-        })
-        cols_definition = generate_columns_definition(BenchmarkType.PE)
-        system_prompts = system_prompts_template.substitute({
-            "columns_definition": cols_definition
-        })
-        msg, usage = client.create(system_prompts,user_message)
-        write_LLM_score(
-            output_fn=ensured_result_path,
-            model=LLModelType.GEMINI15.value,
-            pmid=id,
-            score=msg,
-            token_usage = usage,
-        )
+    run_llm_benchmark(
+        dataset=prepared_dataset,
+        benchmark_type=BenchmarkType.PE,
+        model=LLModelType.GEMINI15,
+        result_file=ensured_result_path,
+        client=client,
+    )
